vehicle-dataset-generator/requirements.txt
# Core packages
openai>=1.97.0
python-dotenv>=1.0.0
requests>=2.32.0
Pillow>=10.0.0

# Web interface (alternative to tkinter)
Flask>=3.0.0

# Object detection for vehicle bounding boxes
ultralytics>=8.0.0
opencv-python>=4.8.0

# Database
mysql-connector-python>=8.0.0
PyMySQL>=1.0.0
SQLAlchemy>=2.0.0

# GUI packages (tkinter는 Python 기본 포함, macOS에서는 별도 설치 필요)
# brew install python-tk

# Optional for advanced features
# bitsandbytes>=0.41.0  # Uncomment if using quantization
